+++ sample/shared/src/iosMain/kotlin/com/sample/shared/Screens.kt
@@ -11,6 +11,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.graphics.Color
 import com.github.guilhe.kmp.composeuiviewcontroller.ComposeUIViewController
 import com.github.guilhe.kmp.composeuiviewcontroller.ComposeUIViewControllerState
+import com.sample.models.ScreenStateExternal
 import platform.Foundation.NSDate
 import platform.Foundation.timeIntervalSince1970
 
@@ -22,7 +23,7 @@ private fun getCurrentMillis(): Long = NSDate().timeIntervalSince1970.toLong() *
 
 @ComposeUIViewController
 @Composable
-internal fun GradientScreen(@ComposeUIViewControllerState state: ScreenState, randomize: (Long) -> Unit) {
+internal fun GradientScreen(@ComposeUIViewControllerState state: ScreenStateExternal, randomize: (Long) -> Unit) {
     Box(contentAlignment = Alignment.Center) {
         Crossfade(targetState = state) {
             Gradient(it.colors)